move mysql pipelining query tests to TCK and enable them in postgres client
